requirements.txt
Django==4.2.0
djangorestframework==3.14.0
PyJWT==2.6.0
python-decouple==3.8
django-cors-headers==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.5.0
dj-database-url==2.1.0


